pyproject.toml
[project]
name = "can-we-update-mc"
version = "0.1.0"
description = "Discord bot to check if your mods are ready for a Minecraft version update"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cord>=2.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "PIE", "PYI", "RSE", "RET", "SLF", "SIM", "ARG", "PTH", "ERA", "PL", "NPY", "RUF"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
strict_equality = true
ignore_missing_imports = true
disable_error_code = [
    "arg-type",
    "return-value",
    "attr-defined",
    "name-defined",
    "type-var",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "types-pytz>=2026.1.1.20260304",
]
